Removed the signup app, instead made sign up and login modals on the homepage.
