+++ schedule_lessons/home/forms.py
@@ -1,0 +1,23 @@
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.models import User
+from django import forms
+
+
+class MyRegistrationForm(UserCreationForm):
+    first_name = forms.CharField(max_length=30, widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'First Name'}))
+    last_name = forms.CharField(max_length=30, widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'Surname'}))
+    CHOICES=[('tutor','Tutor'), ('client','Student')]
+    tutor_or_student = forms.ChoiceField(choices=CHOICES, widget=forms.RadioSelect)
+
+    class Meta:
+        model = User
+        fields = ('email', 'username', 'first_name', 'last_name', 'password1', 'password2', 'tutor_or_student')
+
+    def save(self, commit=True):
+        user = super(MyRegistrationForm, self).save(commit=False)
+        user.first_name = self.cleaned_data['first_name']
+        user.last_name = self.cleaned_data['last_name']
+
+        if commit:
+            user.save()
+        return user
