+++ schedule_lessons/home/static/home/css/index.css
@@ -2,49 +2,25 @@ html, body {
   height: 100%;
   width: 100%;
 }
-/*
-.navbar {
-  background-color: #FEDC32;
-  padding: 1% 0;
-  font-size: 1.2em;
-  margin-bottom: 0px;
-  border: 0px;
-}
-
-.navbar-brand {
-  min-height: 55px;
-  padding: 0 15px 5px;
-}
-
-#logo {
-  width: 65%;
-}
-
-.collapsed {
-  margin-top: 5%;
-}
-
-#bs-example-navbar-collapse-1 {
-  padding: 2% 1% 0 3%;
-}
-
-
-.navbar-default .navbar-nav li a {
-  color: black;
-}
-.navbar-default .navbar-nav>.active>a {
-  color: black;
-  background-color: #edc91a;
-}
-
-.navbar-default .navbar-nav li a:hover, .navbar-default .navbar-nav>.active>a {
-  background-color: #edc91a;
-  color: #FFF;
-} */
 
 .nav-link {
   font-size: 20px;
 }
+
+.errorlist {
+  list-style-type: none;
+  color: red;
+}
+
+#id_tutor_or_student {
+  list-style-type: none;
+  margin-left: 105px;
+}
+
+#id_tutor_or_student>li:nth-child(2) {
+  margin-left: 30px;
+}
+
 
 #home {
   background: url("../images/background.jpg") no-repeat 0% 0% fixed;
