+++ schedule_lessons/schedule_lessons/urls.py
@@ -22,8 +22,6 @@ from django.views.generic.base import RedirectView
 urlpatterns = [
     path('', include('home.urls')),
     path('home/', include('dashboard.urls')),
-    path('signup/', include('signup.urls')),
-    path('login/', login, {'template_name': 'signup/login.html'}),
     path('logout/', logout, name='logout'),
     path('admin/', admin.site.urls),
 ]
