--- schedule_lessons/signup/admin.py
@@ -1,7 +1,0 @@
-from django.contrib import admin
-from .models import Profile
-
-# Register your models here.
-
-
-admin.site.register(Profile)
