--- schedule_lessons/signup/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class SignupConfig(AppConfig):
-    name = 'signup'
