--- schedule_lessons/signup/static/css/signup.css
@@ -1,101 +1,0 @@
-@font-face { font-family: LightFont; src: url('TheLightFont.otf'); }
-@font-face { font-family: Feeding; src: url('Feeding_a_Moment.otf'); }
-@font-face { font-family: Comforta; src: url('Comfortaa-Regular.ttf'); }
-
-
-body {
-    height: auto;
-    box-sizing: border-box;
-}
-
-.header {
-    margin: 0 auto 0 auto;
-    width: 100%;
-    text-align: center;
-}
-
-.header {
-  height: auto;
-  width: 100%;
-}
-.image {
-  width: auto;
-  height: 22%;
-  border: 3px solid black;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-h3 {
-  font-size: 40px;
-  text-align: center;
-  font-family: LightFont;
-  margin: 0 auto;
-}
-
-.content {
-    box-sizing: border-box;
-    margin: 0 auto;
-}
-
-.auth-form {
-    border: 3px solid black;
-    margin: 1.5em auto;
-    float: none;
-    width: 78%;
-    padding: 10px;
-}
-
-/* Signup Button */
-#auth-button {
-    background-color: #ffce0c;
-    width: 200px;
-    border: 3px solid black;
-    border-radius: 5%;
-    color: black;
-}
-
-.auth {
-    margin: 3em auto 2em auto;
-    text-align: center;
-}
-
-
-.form-control {
-    margin-bottom: 1.5em;
-    border: 3px solid black;
-}
-
-.container {
-    width: 500px;
-    height: 600px;
-    margin: 1.5em auto;
-}
-
-.or {
-    text-align: center;
-}
-
-.signup-links {
-  text-align: center;
-  align-items: center;
-    margin: 0 auto;
-}
-
-a.btn {
-    font-size: 16px;
-    text-align: center;
-}
-
-.col-sm-6 {
-  text-align: center;
-  align-items: center;
-}
-
-p {
-    font-weight: bold;
-}
-
-.errorlist, li {
-  color: red;
-}
