--- schedule_lessons/signup/templates/signup/base.html
@@ -1,22 +1,0 @@
-{% load static %}
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
-    
-    {% block css %}{% endblock %}
-  </head>
-  <body class="mai-splash-screen">
-    <div class="mai-wrapper mai-sign-up">
-      <div class="content container">
-          
-        {% block content %}{% endblock %}
-
-      </div>
-    </div>
-  </body>
-  {% block js %}{% endblock %}
-</html>
