--- schedule_lessons/signup/templates/signup/login.html
@@ -1,50 +1,0 @@
-{% extends "signup/base.html" %}
-{% load static %}
-{% block css %}
-    <link rel="stylesheet" type="text/css" href="{% static "css/login.css" %}">
-{% endblock %}
-
-  {% block content %}
-
-
-    <div class="header">
-      <img class="image" src="{% static "images/logo.png" %}" alt="Scheduler">
-
-    </div>
-    <div class="auth-form">
-      <h3>Log In</h3>
-      <form method="post">
-        {% csrf_token %}
-        {% for field in form %}
-          <p>
-            {{ field.label_tag }}<br>
-            {{ field }}
-          </p>
-        {% endfor %}
-        {% if form.errors %}
-        <ul>
-          <li class="error_class">Not a valid username/password combination. Try again.</li>
-        </ul>
-        {% endif %}
-        <div class="auth">
-          <button id="auth-button" class="btn" type="submit">Log In</button>
-        </div>
-      </form>
-      <div class="or">
-        <p > Or Sign Up As: </p>
-
-      </div>
-      <div class="row signup-links">
-        <div class="col-sm-6">
-            <a href="/signup/client" class="btn">Student</a>
-        </div>
-        <div class="col-sm-6">
-            <a href="/signup/tutor" class="btn">Tutor</a>
-        </div>
-      </div>
-    </div>
-  {% endblock %}
-
-  {% block js %}
-    <script src="{% static "js/login.js" %}" type="text/javascript"></script>
-  {% endblock %}
