--- schedule_lessons/signup/templates/signup/signup.html
@@ -1,46 +1,0 @@
-{% extends "signup/base.html" %}
-{% load static %}
-
-{% block css %}
-  <link rel="stylesheet" type="text/css" href="{% static "css/signup.css" %}">
-{% endblock %}
-
-  {% block content %}
-  <div class="header">
-    <img class="image" src="{% static "images/logo.png" %}" alt="Scheduler">
-
-    </div>
-    <div class="auth-form">
-
-      <h3>Sign Up</h3>
-      <form method="post" action="{{ user_type }}">
-        {% csrf_token %}
-        {% for field in form %}
-          <p>
-            {{ field.label_tag }}<br>
-            {{ field }}
-            {% for error_field, message in errors %}
-              {% if error_field == field.name %}
-                {{ message }}
-              {% endif %}
-            {% endfor %}
-          </p>
-        {% endfor %}
-
-        <div class="auth">
-          <button id="auth-button" class="btn" type="submit">Sign up</button>
-        </div>
-        <div class="or">
-            <p> OR </p>
-          </div>
-          <div class="row signup-links">
-              <a href="/login" class="btn">Login</a>
-          </div>
-        </div>
-      </form>
-    </div>
-  {% endblock %}
-
-  {% block js %}
-    <script src="{% static "js/signup.js" %}" type="text/javascript"></script>
-  {% endblock %}
