+++ schedule_lessons/static/js/login.js
