--- schedule_lessons/templates/about.html
@@ -1,70 +1,0 @@
-<!DOCTYPE html>
-{% load static %}
-<html>
-  <head>
-    <title>Home | Schedulearn</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.1/css/bootstrap.min.css" integrity="sha384-WskhaSGFgHYWDcbwN70/dfYBj47jz9qbsMId/iRN3ewGhXQFZCSftd1LZCfmhktB" crossorigin="anonymous">
-  <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.0.13/css/all.css" integrity="sha384-DNOHZ68U8hZfKXOrtjWvjxusGo9WQnrNx2sqG0tfsghAvtVlRW3tvkXWZh58N9jp" crossorigin="anonymous">
-  <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
-<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.1/js/bootstrap.min.js" integrity="sha384-smHYKdLADwkXOn1EmN1qk/HfnUcbVRZyYmZ4qpPea6sjB/pTJ0euyQp0Mk8ck+5T" crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="{% static 'home/css/index.css' %}">
-  </head>
-  <body>
-    <nav class="navbar navbar-expand-lg navbar-light" style="background-color: #FEDC32;">
-      <a class="navbar-brand" href="#">
-        <img src="{% static 'home/images/logo_transparent.png'%}" alt="">
-      </a>
-      <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-      </button>
-
-      <div class="collapse navbar-collapse" id="navbarSupportedContent">
-        <ul class="navbar-nav mr-auto">
-          <li class="nav-item">
-            <a class="nav-link" href="/">Home</a>
-          </li>
-          <li class="nav-item active">
-            <a class="nav-link" href="#">About Schedulearn<span class="sr-only">(current)</span></a>
-          </li>
-        </ul>
-        <ul class="navbar-nav ml-auto">
-          <li class="nav-item">
-            <a class="nav-link" href="/login/">Log In</a>
-          </li>
-        </ul>
-      </div>
-    </nav>
-
-    <div id="about-schedulearn" class="padding">
-      <div class="container">
-        <div class="row">
-          <div class="col-sm-6">
-            <img style="width:90%" src="{% static 'home/images/demo.png' %}">
-          </div>
-          <div class="col-sm-6 text-center">
-            <h2>About Schedulearn</h2>
-            <p class="lead">Schedulearn is all about connecting students and professors in a way that hasn't been done before. The way Schedulearn does that is by giving tutors the ability to give students assignments, schedule lessons and much more.</p>
-            <p class="lead">Students get access to features that enhance their learning experience such as chatrooms with other students, the ability to schedule lessons with their teacher and more.</p>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <footer class="container-fluid text-center">
-      <div class="row">
-        <div class="col-sm-6">
-          <h3>Contact Us</h3><br>
-          <h4>Our contact information is here.</h4>
-        </div>
-        <div class="col-sm-6">
-          <h3>Connect</h3>
-          <a href="https://github.com/sdevalapurkar/schedule-my-lessons" class="fab fa-github"></a>
-        </div>
-
-      </div>
-    </footer>
-
-  </body>
-</html>
